Add `datafusion-functions-array` crate
