+++ datafusion/proto/src/logical_plan/from_proto.rs
@@ -52,11 +52,10 @@ use datafusion_expr::{
     array_intersect, array_length, array_ndims, array_pop_back, array_pop_front,
     array_position, array_positions, array_prepend, array_remove, array_remove_all,
     array_remove_n, array_repeat, array_replace, array_replace_all, array_replace_n,
-    array_resize, array_slice, array_sort, array_to_string, array_union, arrow_typeof,
-    ascii, asin, asinh, atan, atan2, atanh, bit_length, btrim, cardinality, cbrt, ceil,
-    character_length, chr, coalesce, concat_expr, concat_ws_expr, cos, cosh, cot,
-    current_date, current_time, date_bin, date_part, date_trunc, degrees, digest,
-    ends_with, exp,
+    array_resize, array_slice, array_sort, array_union, arrow_typeof, ascii, asin, asinh,
+    atan, atan2, atanh, bit_length, btrim, cardinality, cbrt, ceil, character_length,
+    chr, coalesce, concat_expr, concat_ws_expr, cos, cosh, cot, current_date,
+    current_time, date_bin, date_part, date_trunc, degrees, digest, ends_with, exp,
     expr::{self, InList, Sort, WindowFunction},
     factorial, find_in_set, flatten, floor, from_unixtime, gcd, gen_range, initcap,
     instr, isnan, iszero, lcm, left, levenshtein, ln, log, log10, log2,
@@ -507,7 +506,6 @@ impl From<&protobuf::ScalarFunction> for BuiltinScalarFunction {
             ScalarFunction::ArrayReplaceAll => Self::ArrayReplaceAll,
             ScalarFunction::ArrayReverse => Self::ArrayReverse,
             ScalarFunction::ArraySlice => Self::ArraySlice,
-            ScalarFunction::ArrayToString => Self::ArrayToString,
             ScalarFunction::ArrayIntersect => Self::ArrayIntersect,
             ScalarFunction::ArrayUnion => Self::ArrayUnion,
             ScalarFunction::ArrayResize => Self::ArrayResize,
@@ -1461,10 +1459,6 @@ pub fn parse_expr(
                     parse_expr(&args[1], registry)?,
                     parse_expr(&args[2], registry)?,
                     parse_expr(&args[3], registry)?,
-                )),
-                ScalarFunction::ArrayToString => Ok(array_to_string(
-                    parse_expr(&args[0], registry)?,
-                    parse_expr(&args[1], registry)?,
                 )),
                 ScalarFunction::Range => Ok(gen_range(
                     args.to_owned()
